requirements.txt
<<<<<<< HEAD
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
boto3==1.37.31
botocore==1.37.31
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
frozenlist==1.5.0
fsspec==2025.3.2
gradio==5.25.2
gradio_client==1.8.0
greenlet==3.2.0
groovy==0.1.2
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.54
langchain-text-splitters==0.3.8
langsmith==0.3.32
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.4.2
mypy-extensions==1.0.0
numpy==2.2.4
openai==1.72.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
pluggy==1.5.0
propcache==0.3.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.25.5
PyMuPDFb==1.24.10
pypdfium2==4.30.1
pytest==8.3.5
pytest-asyncio==0.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
ruff==0.11.6
s3transfer==0.11.4
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
stripe==11.5.0
tenacity==9.1.2
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.19.0
youtube-transcript-api==1.0.3
zstandard==0.23.0
=======
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
frozenlist==1.5.0
fsspec==2025.3.2
gradio==5.25.2
gradio_client==1.8.0
greenlet==3.2.0
groovy==0.1.2
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.54
langchain-text-splitters==0.3.8
langsmith==0.3.32
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.4.2
mypy-extensions==1.0.0
numpy==2.2.4
openai==1.72.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pinecone
pinecone-plugin-interface==0.0.7
pluggy==1.5.0
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.25.5
PyMuPDFb==1.24.10
pytest==8.3.5
pytest-asyncio==0.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
ruff==0.11.6
s3transfer==0.11.4
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
stripe==11.5.0
tenacity==9.1.2
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.2.0
Werkzeug==3.0.1
yarl==1.9.4
>>>>>>> a46afa3 (initial commit)
